staticfiles/requirements.txt
antiorm==1.2.1
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
db==0.1.1
db-sqlite3==0.0.1
Django==5.0.6
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-js-asset==2.2.0
html5lib==1.1
HTMLParser==0.0.2
idna==3.7
Jinja2==3.1.4
lxml==5.2.1
mailmerge==2.2.3
Markdown==3.6
MarkupSafe==2.1.5
pillow==10.3.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
